+++ src/types/GeneratorState.ts
@@ -11,6 +11,11 @@ export enum Stages {
     Complete = 'Complete',
 }
 
+export interface ResourceSourceState {
+    syncTime?: number;
+    syncHash?: string;
+}
+
 export interface GeneratorState {
     stage?: string;
     lastUpdated?: number;
@@ -19,4 +24,5 @@ export interface GeneratorState {
     processedStages?: number;
     possibleStagesRemaining?: number;
     childResources?: number;
+    resourceStatus?: ResourceSourceState[];
 }